update linear-gradient for Safari Compatibility
In Safari, Linear-Gradients, when colour interpolating, let's say, from white to another colour with a hue of 0 and/or a lightness of zero with an alpha of 0, will in turn cause the latter color (hue 0, alpha 0) to interpolate from the previous color, black to transparent. This is different in other browsers such as chrome. In response to this I've updated the lightness to 100% so that the gradient will interpolate from white rather than black.